Ebola 1976 outbreak case data

An example epidemic outbreak dataset for use with the cfr package. This dataset comes from the first Ebola outbreak in Zaire in 1976 as analysed in Camacho et al. (2014).

Usage

ebola1976

Format

ebola1976

A <data.frame> with 73 rows and 3 columns:

date

Calendar date

cases

Number of cases reported

deaths

Number of deaths reported

Source

doi:10.1016/j.epidem.2014.09.003

References

Camacho, A., Kucharski, A. J., Funk, S., Breman, J., Piot, P., & Edmunds, W. J. (2014). Potential for large outbreaks of Ebola virus disease. Epidemics, 9, 70–78. doi:10.1016/j.epidem.2014.09.003