Estimate known outcomes from case and death time-series data

Estimates the expected number of individuals with known outcomes from a case and death time series of data of an outbreak, up to the time point supplied. Calculates the daily new number of known outcomes.

Uses the probability mass function representing the delay between case detection and death, typically approximated by a symptom onset to death distribution from the literature for the disease in question.

Usage

known_outcomes(data, epidist = NULL)

Arguments

data

A data.frame containing the outbreak data. A daily time series with dates or some other absolute indicator of time (e.g. epiday/epiweek) and the numbers of new cases at each time point. This function does not require data on daily deaths, but this column (and any others) will be retained if present.

epidist

The delay distribution used, in the form of an epiparameter::epidist() object. This is used to obtain a probability mass function parameterised by time; i.e. \(f(t)\) which gives the probability a case has a known outcomes (i.e. death) at time \(t\), parameterised with disease-specific parameters before it is supplied here. A typical example would be a symptom onset to death delay distribution.

Value

A data.frame with the columns in data, and with two additional columns,

"known_outcomes" for the total number of known outcomes on that day of the outbreak, and

"u_t" for the under-reporting factor.

Examples

# Load Ebola 1976 outbreak data
data("ebola1976")

# read epidist for EVD onset to death from {epiparameter}
onset_to_death_ebola <- epiparameter::epidist_db(
  disease = "Ebola Virus Disease",
  epi_dist = "onset_to_death",
  author = "Barry_etal"
)
#> Using Barry et al. (2018) <10.1016/S0140-6736(18)31387-4> PMID: 30047375. 
#> To retrieve the short citation use the 'get_citation' function

# examine the first few rows of the output
head(
  known_outcomes(data = ebola1976, onset_to_death_ebola)
)
#>         date cases deaths known_outcomes        u_t
#> 1 1976-08-25     1      0     0.00000000 0.00000000
#> 2 1976-08-26     0      0     0.03316057 0.03316057
#> 3 1976-08-27     0      0     0.06482994 0.09799050
#> 4 1976-08-28     0      0     0.08472567 0.18271617
#> 5 1976-08-29     0      0     0.09389465 0.27661082
#> 6 1976-08-30     0      0     0.09506629 0.37167711