Estimate a severity measure that varies over time

Calculates how the severity of a disease changes over time while optionally correcting for reporting delays using an epidemiological delay distribution of the time between symptom onset and death (onset-to-death).

Usage

cfr_time_varying(
  data,
  epidist = NULL,
  burn_in = get_default_burn_in(epidist),
  smoothing_window = NULL
)

Arguments

data

A <data.frame> containing the outbreak data. A daily time series with dates or some other absolute indicator of time (e.g. epiday or epiweek) and the numbers of new cases and new deaths at each time point. Note that the required columns are "date" (for the date), "cases" (for the number of reported cases), and "deaths" (for the number of reported deaths) on each day of the outbreak.

Note that the <data.frame> is required to have an unbroken sequence of dates with no missing dates in between. The "date" column must be of class Date (see as.Date()).

Note also that the total number of cases must be greater than the total number of reported deaths.

epidist

An optional argument for the delay distribution used, in the form of an <epidist> object. Defaults to NULL, in which case no delay correction is applied. Passing an <epidist> automatically applies delay correction in the severity estimation.

burn_in

A single integer value for the number of time-points (typically days) to disregard at the start of the time-series, if a burn-in period is desired.

When an <epidist> is provided and delay correction is requested, the default value is set to the mean of the <epidist>.

Defaults to 7 if no <epidist> is provided, or if the <epidist> is not parameterised. This is a sensible default value that disregards the first week of cases and deaths, assuming daily data.

To consider all case data including the start of the time-series, set this argument to 0.

smoothing_window

An odd number determining the smoothing window size to use when smoothing the case and death time-series, using a rolling median procedure (as the k argument to stats::runmed()) before calculating the time-varying severity.

The default behaviour is to apply no smoothing. The minimum value of this argument is 1.

Value

A <data.frame> containing the maximum likelihood estimate estimate and 95% confidence interval of the corrected severity.

Details: Adjusting for delays between two time series

This function estimates the number of cases which have a known outcome over time, following Nishiura et al. (2009). The function calculates a quantity \(k_t\) for each day within the input data, which represents the number of cases with a known outcome, on day \(t\). \(k_t\) is calculated in the following way: $$k_t = \sum_{j = 0}^t c_t f_{j - t}$$

We then assume that the severity measure, for example CFR, of interest is binomially distributed, in the following way:

$$d_t \sim {\sf Binomial}(k_t, \theta_t)$$

We use maximum likelihood estimation to determine the value of \(\theta_t\) for each \(t\), where \(\theta\) represents the severity measure of interest.

The epidemiological delay distribution passed to epidist is used to obtain a probability mass function parameterised by time; i.e. \(f(t)\) which gives the probability a case has a known outcomes (usually, death) at time \(t\), parameterised with disease-specific parameters before it is supplied here.

References

Nishiura, H., Klinkenberg, D., Roberts, M., & Heesterbeek, J. A. P. (2009). Early Epidemiological Assessment of the Virulence of Emerging Infectious Diseases: A Case Study of an Influenza Pandemic. PLOS ONE, 4(8), e6852. doi:10.1371/journal.pone.0006852

Examples

# get data pre-loaded with the package
data("covid_data")
df_covid_uk <- covid_data[covid_data$country == "United Kingdom", ]

# load epidist object from {epiparameter}
onset_to_death_covid <- epiparameter::epidist_db(
  disease = "COVID-19",
  epi_dist = "onset_to_death",
  author = "Linton_etal",
  single_epidist = TRUE
)
#> Using list(author = list(list(given = NULL, family = "Linton", role = NULL, email = NULL, comment = NULL), list(given = NULL, family = "etal", role = NULL, email = NULL, comment = NULL)), year = "2020", title = "Incubation Period and Other Epidemiological Characteristics of 2019 Novel Coronavirus Infections with Right Truncation: A Statistical Analysis of Publicly Available Case Data", journal = "Journal of Clinical Medicine", doi = "10.3390/jcm9020538", pmid = "32079150"). 
#> To retrieve the short citation use the 'get_citation' function

# estimate time varying severity without correcting for delays
cfr_time_varying <- cfr_time_varying(
  data = df_covid_uk,
  burn_in = 7L
)
# View
tail(cfr_time_varying)
#>             date        country cases deaths severity_mean severity_low
#> 20690 2022-12-26 United Kingdom  4132    144    0.03484995   0.02946797
#> 20709 2022-12-27 United Kingdom  3694    131    0.03546291   0.02973354
#> 20728 2022-12-28 United Kingdom  4950    128    0.02585859   0.02161760
#> 20747 2022-12-29 United Kingdom  6605    142    0.02149886   0.01813784
#> 20766 2022-12-30 United Kingdom  8419    134    0.01591638   0.01335220
#> 20785 2022-12-31 United Kingdom  7718    132    0.01710288   0.01432899
#>       severity_high
#> 20690    0.04090181
#> 20709    0.04194157
#> 20728    0.03067046
#> 20747    0.02529105
#> 20766    0.01882303
#> 20785    0.02024949

# estimate time varying severity while correcting for delays
cfr_time_varying <- cfr_time_varying(
  data = df_covid_uk,
  epidist = onset_to_death_covid,
  burn_in = 7L
)
tail(cfr_time_varying)
#>             date        country cases deaths severity_mean severity_low
#> 20690 2022-12-26 United Kingdom  4132    144    0.03471553   0.02935401
#> 20709 2022-12-27 United Kingdom  3694    131    0.03065762   0.02569476
#> 20728 2022-12-28 United Kingdom  4950    128    0.02893963   0.02419931
#> 20747 2022-12-29 United Kingdom  6605    142    0.03093008   0.02611344
#> 20766 2022-12-30 United Kingdom  8419    134    0.02815718   0.02364360
#> 20785 2022-12-31 United Kingdom  7718    132    0.02684018   0.02250428
#>       severity_high
#> 20690    0.04074453
#> 20709    0.03627492
#> 20728    0.03431472
#> 20747    0.03635481
#> 20766    0.03326130
#> 20785    0.03174924